Add login view for tablet
