+++ lib/pages/tablet/login_form.dart
@@ -8,6 +8,7 @@ import 'package:k_front/widgets/common/app_bar.dart';
 import '../../models/logged_user.dart';
 import '../../themes/theme_data.dart';
 import '../../routes/dashboard.dart';
+import '../../widgets/common/o_auth_buttons.dart';
 import '../../widgets/desktop/background_index.dart';
 
 class TabletLoginForm extends StatefulWidget {
@@ -20,7 +21,6 @@ class TabletLoginForm extends StatefulWidget {
 class TabletLoginFormState extends State<TabletLoginForm> {
   final GlobalKey<FormState> _formKey = GlobalKey<FormState>();
   final String title = "kCal Control";
-  final FocusScopeNode _focusNode = FocusScopeNode();
   bool _obscureText = true;
   BuildContext? _navigationContext;
   final _newLoggedUser = LoggedUser(
@@ -86,38 +86,8 @@ class TabletLoginFormState extends State<TabletLoginForm> {
     return false;
   }
 
-  void _handleKeyEvent() {
-    if (_focusNode.hasFocus) {
-      HardwareKeyboard.instance.addHandler(_handleRawKeyEvent);
-    } else {
-      HardwareKeyboard.instance.removeHandler(_handleRawKeyEvent);
-    }
-  }
-
-  bool _handleRawKeyEvent(KeyEvent event) {
-    if (event is KeyboardKey && event.logicalKey == LogicalKeyboardKey.enter) {
-      _login();
-      return true;
-    }
-    return false;
-  }
-
   bool handleCheckbox(bool value) {
     return !value;
-  }
-
-  @override
-  void initState() {
-    super.initState();
-    _focusNode.requestFocus();
-    _focusNode.addListener(_handleKeyEvent);
-  }
-
-  @override
-  void dispose() {
-    _focusNode.removeListener(_handleKeyEvent);
-    HardwareKeyboard.instance.removeHandler(_handleRawKeyEvent);
-    super.dispose();
   }
 
   @override
@@ -132,161 +102,161 @@ class TabletLoginFormState extends State<TabletLoginForm> {
         body: Stack(
           children: [
             const BackgroundScreen(),
-            Row(
-              mainAxisAlignment: MainAxisAlignment.center,
-              children: <Widget>[
-                TextButton(
-                  onPressed: () {},
-                  child: Text(
-                      style: Theme.of(context).textTheme.labelSmall,
-                      'Terms and Conditions'),
-                ),
-                const Text('|'),
-                TextButton(
-                  onPressed: () {},
-                  child: Text(
-                      style: Theme.of(context).textTheme.labelSmall,
-                      'Privacy Policy'),
-                ),
-              ],
-            ),
-            FocusScope(
-              node: _focusNode,
-              child: Center(
-                child: Form(
-                    key: _formKey,
-                    child: Container(
+            Center(
+              child: SingleChildScrollView(
+                child: Column(
+                  mainAxisAlignment: MainAxisAlignment.center,
+                  crossAxisAlignment: CrossAxisAlignment.center,
+                  children: [
+                    Container(
                       width: 900,
                       height: 450,
                       decoration: kContainerDecoration.copyWith(
-                        color: Theme.of(context).canvasColor,
-                      ),
+                          color: Theme.of(context).canvasColor,
+                        ),
                       padding: const EdgeInsets.symmetric(
-                          horizontal: 24.0, vertical: 30),
-                      child: Column(
-                          mainAxisAlignment: MainAxisAlignment.spaceAround,
-                          crossAxisAlignment: CrossAxisAlignment.center,
-                          children: <Widget>[
-                            const Text(
-                              'Welcome Back!',
-                              style: TextStyle(
-                                fontSize: 20,
-                                fontWeight: FontWeight.bold,
-                              ),
-                            ),
-                            Row(
-                              children: [
-                                Expanded(
-                                  child: Column(
+                            horizontal: 40, vertical: 30),
+                      child: Center(
+                        child: Form(
+                          key: _formKey,
+                          child: Column(
+                                mainAxisAlignment: MainAxisAlignment.spaceAround,
+                                crossAxisAlignment: CrossAxisAlignment.center,
+                                children: <Widget>[
+                                  const Text(
+                                    'Welcome Back!',
+                                    style: TextStyle(
+                                      fontSize: 20,
+                                      fontWeight: FontWeight.bold,
+                                    ),
+                                  ),
+                                  Row(
                                     children: [
-                                      const SizedBox(height: 20),
-                                      buildTextField(
-                                        hintText: 'Email or username',
-                                        icon: Icons.person,
-                                        onSave: (value) {
-                                          _newLoggedUser.username = value!;
-                                        },
+                                      Expanded(
+                                        child: Column(
+                                          children: [
+                                            const SizedBox(height: 20),
+                                            buildTextField(
+                                              hintText: 'Email or username',
+                                              icon: Icons.person,
+                                              onSave: (value) {
+                                                _newLoggedUser.username = value!;
+                                              },
+                                            ),
+                                            const SizedBox(height: 20),
+                                            buildTextField(
+                                              hintText: 'Password',
+                                              icon: Icons.lock,
+                                              obscureText: true,
+                                              onSave: (value) {
+                                                _newLoggedUser.password = value!;
+                                              },
+                                            ),
+                                            const SizedBox(height: 10),
+                                            Row(
+                                              children: <Widget>[
+                                                Checkbox(
+                                                  value: _rememberMe,
+                                                  onChanged: (value) {
+                                                    setState(() {
+                                                      _rememberMe = value!;
+                                                    });
+                                                  },
+                                                ),
+                                                const Text('Remember me'),
+                                              ],
+                                            ),
+                                            const SizedBox(height: 10),
+                                            ElevatedButton(
+                                              onPressed: () async {
+                                                await _login()
+                                                    ? Navigator.push(
+                                                        _navigationContext!,
+                                                        MaterialPageRoute(
+                                                            builder: (context) =>
+                                                                const Dashboard()))
+                                                    : null;
+                                              },
+                                              style: ElevatedButton.styleFrom(
+                                                backgroundColor:
+                                                    Theme.of(context).splashColor,
+                                                padding: const EdgeInsets.symmetric(
+                                                    horizontal: 40, vertical: 5),
+                                                shape: RoundedRectangleBorder(
+                                                  borderRadius:
+                                                      BorderRadius.circular(10.0),
+                                                ),
+                                              ),
+                                              child: const Text('Sign In'),
+                                            ),
+                                            TextButton(
+                                              onPressed: () {},
+                                              child: const Text('Forgot password?'),
+                                            ),
+                                          ],
+                                        ),
                                       ),
-                                      const SizedBox(height: 20),
-                                      buildTextField(
-                                        hintText: 'Password',
-                                        icon: Icons.lock,
-                                        obscureText: true,
-                                        onSave: (value) {
-                                          _newLoggedUser.password = value!;
-                                        },
-                                      ),
-                                      const SizedBox(height: 10),
-                                      Row(
-                                        children: <Widget>[
-                                          Checkbox(
-                                            value: _rememberMe,
-                                            onChanged: (value) {
-                                              setState(() {
-                                                _rememberMe = value!;
-                                              });
-                                            },
-                                          ),
-                                          const Text('Remember me'),
-                                        ],
-                                      ),
-                                      ElevatedButton(
-                                        onPressed: () async {
-                                          await _login()
-                                              ? Navigator.push(
-                                                  _navigationContext!,
-                                                  MaterialPageRoute(
-                                                      builder: (context) =>
-                                                          const Dashboard()))
-                                              : null;
-                                        },
-                                        style: ElevatedButton.styleFrom(
-                                          backgroundColor:
-                                              Theme.of(context).splashColor,
-                                          padding: const EdgeInsets.symmetric(
-                                              horizontal: 40, vertical: 5),
-                                          shape: RoundedRectangleBorder(
-                                            borderRadius:
-                                                BorderRadius.circular(10.0),
+                                      Padding(
+                                        padding: const EdgeInsets.all(24.0),
+                                        child: SizedBox(
+                                          height: 300,
+                                          child: VerticalDivider(
+                                            color: Theme.of(context).disabledColor,
+                                            thickness: 1.5,
+                                            width: 1,
+                                            endIndent: 20,
+                                            indent: 20,
                                           ),
                                         ),
-                                        child: const Text('Sign In'),
                                       ),
-                                      TextButton(
-                                        onPressed: () {},
-                                        child: const Text('Forgot password?'),
-                                      ),
+                                      Column(
+                                          mainAxisAlignment: MainAxisAlignment.center,
+                                          crossAxisAlignment: CrossAxisAlignment.center,
+                                          children: [
+                                            SizedBox(
+                                                width: 200,
+                                                child: oAuth2Buttons(
+                                                    text: 'Sign up with Google',
+                                                    icon: Icons.g_mobiledata)),
+                                            const SizedBox(height: 15),
+                                            SizedBox(
+                                                width: 200,
+                                                child: oAuth2Buttons(
+                                                    text: 'Sign up with Microsoft',
+                                                    icon: Icons.window)),
+                                            const SizedBox(height: 15),
+                                            SizedBox(
+                                                width: 200,
+                                                child: oAuth2Buttons(
+                                                    text: 'Sign up with Apple',
+                                                    icon: Icons.apple)),
+                                          ])
                                     ],
                                   ),
-                                ),
-                                Padding(
-                                  padding: const EdgeInsets.all(24.0),
-                                  child: SizedBox(
-                                    height: 200,
-                                    child: VerticalDivider(
-                                      color: Theme.of(context).disabledColor,
-                                      thickness: 1,
-                                      width: 1,
-                                      endIndent: 20,
-                                      indent: 20,
-                                    ),
-                                  ),
-                                ),
-                                Expanded(
-                                    child: Column(
-                                  children: [
-                                    ElevatedButton(
-                                      onPressed: () async {
-                                        await _login()
-                                            ? Navigator.push(
-                                                _navigationContext!,
-                                                MaterialPageRoute(
-                                                    builder: (context) =>
-                                                        const Dashboard()))
-                                            : null;
-                                      },
-                                      style: ElevatedButton.styleFrom(
-                                        backgroundColor:
-                                            Theme.of(context).splashColor,
-                                        padding: const EdgeInsets.symmetric(
-                                            horizontal: 40, vertical: 5),
-                                        shape: RoundedRectangleBorder(
-                                          borderRadius:
-                                              BorderRadius.circular(10.0),
-                                        ),
-                                      ),
-                                      child: const Text('Sign In'),
-                                    ),
-                                    TextButton(
-                                      onPressed: () {},
-                                      child: const Text('Forgot password?'),
-                                    ),
-                                  ],
-                                )),
-                              ],
-                            ),
-                          ]),
-                    )),
+                                ]),
+                        ),
+                      ),
+                    ),
+                    Row(
+                      mainAxisAlignment: MainAxisAlignment.center,
+                      children: <Widget>[
+                        TextButton(
+                          onPressed: () {},
+                          child: Text(
+                              style: Theme.of(context).textTheme.labelSmall,
+                              'Terms and Conditions'),
+                        ),
+                        const Text('|'),
+                        TextButton(
+                          onPressed: () {},
+                          child: Text(
+                              style: Theme.of(context).textTheme.labelSmall,
+                              'Privacy Policy'),
+                        ),
+                      ],
+                    ),
+                  ],
+                ),
               ),
             ),
           ],
