+++ lib/pages/tablet/signup_form.dart
@@ -6,8 +6,9 @@ import 'package:k_front/controller/api_user_controller.dart';
 import '../../controller/impl/api_user_controller_impl.dart';
 import '../../themes/theme_data.dart';
 import '../../widgets/common/app_bar.dart';
+import '../../widgets/common/o_auth_buttons.dart';
 import '../../widgets/desktop/background_index.dart';
-import '../../widgets/phone_selector.dart';
+import '../../widgets/common/phone_selector.dart';
 
 //TODO unify the signup form for all devices
 
@@ -250,26 +251,6 @@ class TabletSignUpPageState extends State<TabletSignUpPage> {
       ]),
       floatingActionButton: themeSelectorButton(context),
       floatingActionButtonLocation: FloatingActionButtonLocation.endFloat,
-    );
-  }
-
-  ElevatedButton oAuth2Buttons({required String text, required IconData icon}) {
-    return ElevatedButton(
-      style: ElevatedButton.styleFrom(
-        padding: const EdgeInsets.symmetric(horizontal: 10, vertical: 15),
-        shape: RoundedRectangleBorder(
-          borderRadius: BorderRadius.circular(10.0),
-        ),
-      ),
-      onPressed: () => {},
-      child: Row(
-        mainAxisAlignment: MainAxisAlignment.start,
-        children: <Widget>[
-          Icon(icon),
-          const SizedBox(width: 10),
-          Text(text),
-        ],
-      ),
     );
   }
