+++ lib/widgets/common/phone_selector.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:intl_phone_number_input/intl_phone_number_input.dart';
 import 'package:provider/provider.dart';
 
-import '../utils/phone_validator.dart';
+import '../../utils/phone_validator.dart';
 
 InternationalPhoneNumberInput internationalPhoneNumberInput(
     PhoneNumber number, BuildContext context, TextEditingController phoneController, newUser) {
